requirements/requirements.txt
fire==0.4.0
ruamel.yaml==0.17.17
GitPython==3.1.29
addict==2.3.0
python-docx==1.1.0
mleco==2.0.1
mlcore.mleco-loader==1.0.1
jinja2==3.1.3
markdown2
weasyprint